// types/contentfulTypes.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface IPenFields {
  /** Title */
  title: string;

  /** Hero */
  hero: Asset;

  /** Post */
  post: Document;

  /** Author */
  author: 'Alex Ramos' | 'Walker Spence';
}

/** A post.  */

export interface IPen extends Entry<IPenFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'pen';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export type CONTENT_TYPE = 'pen';

export type LOCALE_CODE = 'en-US';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'en-US';